cleaned up configuration error feedback
figwheel now prints out the build-id that is being focussed on.  It
also will fail if no such build id exists.  Previously it would default
to the first build that had :optimizations set to :none.
This is confusing behavior.  If you specify a build id that is
misconfigured now leon figwheel will fail and report why.

Also cleaned up and commented code a bit to get ready for writing tests.