Write nREPL port number to .nrepl-port file
When connecting to an nREPL server started by Figwheel via `M-x
cider-connect` you have to enter the hostname and the port of the nREPL
server. The hostname defaults to `localhost`, and the port number is
taken from the `.nrepl-port` file if it exists. This commit writes the
port number taken from the Figwheel options to that file when starting a
nREPL server. This allows you to invoke `M-x cider-connect` and hit
enter 2 times, instead of entering the port number manually.