Use :system-running in figwheel-running?
This change updates `figwheel-running?` in the sidecar repl-api
namespace to use the `{:figwheel-system {:system-running Bool}}` value
instead of the truthy value of the *repl-api-system*.

This is required because when `stop-figwheel!` stops the component it
updates the var with the stopped version (truthy) instead of setting it
back to a nil (falsy). This leads to the case where `figwheel-running?`
will return true after a figwheel component has been started and
stopped.