integrating lein profile merging
Need to keep the validation code out of the lein process and
leinigen-core out of the running process. And doing the validation in a
subprocess instead of the lein process.

This should eliminate all the various booting errors folks had with
figwheel.

This will also allow start-figwheel! to be much smarter about loading
configuration files.