Add support for lein's `:managed-dependencies`
Prior to this commit, figwheel would error out on any projects
that were using lein's `:managed-dependencies` feature to
inject version numbers into the project.  This commit makes sure
that `:managed-dependencies` is one of the included keys selected
from the original project map, and allows projects that use this
key to run properly.