Set Cache-Control: no-cache on responses
Cache-Control: no-cache tells the browser that it may cache the
response, but it must always revalidate the cached response with the
server, before using it. This is equivalent to setting

Cache-Control: max-age=0, must-revalidate

max-age=0 means that the response is always stale, and must-revalidate
means that the client must always revalidate stale resources.

Expires: -1 is not needed, all modern browsers support Cache-Control.
IE8 (and below?) treat 'Cache-Control: no-cache' as 'no-store', meaning
that they will always revalidate the cache. That seems a small price to
pay for the simpler reasoning about caching by only using Cache-Control.

https://developers.google.com/web/fundamentals/performance/optimizing-content-efficiency/http-caching
has good information on the current state of the art for HTTP caching.
https://jakearchibald.com/2016/caching-best-practices/#pattern-2-mutable-content-always-server-revalidated
has a good explanation of exactly the caching pattern that we are
looking to employ.

See also https://stackoverflow.com/a/19938619/826486 for more discussion
on Cache-Control: no-cache.

Based on #464, but fixes merge conflicts, and removes Expires: -1.

Fixes #546.